upload

+++ math/fast_power.h
@@ -13,11 +13,11 @@ vector<bool> dec2bin(int src)
     return ret;
 }
 
-long long binpow(long long a, long long b)
+size_t binpow(size_t a, size_t b)
 {
     if (b == 0)
         return 1;
-    long long res = binpow(a, b / 2);
+    size_t res = binpow(a, b / 2);
     if (b % 2)
         return res * res * a;
     else
